Add accepted file_types to http OPTIONS of the Dataset.archive
Resolves #41

+ added detail routes to OPTIONS method on datasets.  This includes ('submit','approve','unsubmit','unapprove','upload','archive'
+ upload defines required parameter 'attachement' and it's allowed_mime_types.
